Records and test set change
Fixed a bug when the user sets both records=r  and test=t.  In this
case, the user wants to use r instances of data and set aside t% of that
data for testing.  However, it used all r instances for training and set
aside an additional t% of data for testing on top of that, using more
total data than the user requested.  This seems to be the less intuitive
way to do it, although it is still valid to test and train that way.
Now, r instances are used in total for training and testing as that
seems to be the more common use case

Set default slice of URL data to be the first 10 features (for fast
demos).